dinfermodel: don't fail if NC_NETCDF4 is allowed, but NC3 is found
In VTK, there are some files which require the NC3 implementation, but
no longer open under 4.8.0 (it worked under 4.7.4). The code checks to
make sure that certain formats were *not* requested when it is entirely
reasonable that support may be required for other files.

This partially reverts changes made in
59e04ae0712a59688e5ae80dfbe7fc4c295d2826 which is a massive commit which
adds Zarr support but doesn't mention why this specific change was made.