WIP: attribute creation order on/off
Work in progress / Proof of concept:

Add a capability to disable the tracking of attribute creation order.
See #2054 for details.

This PR adds a `NC_NOATTCREORD` define which can be passed int the
`mode` argument to `nc_create`.  If it is present, then the
calls to set the attribute creation order tracking is disabled.
This should only be used for files in which you *know* that the
ordering of the attributes does not matter to *any* potential
readers of this database.