Detect attribute creation order tracking setting
When opening an existing file for NC_WRITE access,
check whether the file was created originally with
attribute creation order tracking disabled and if
so, use that setting for subsequent attribute creation.

Also check the `mode` passed in to the open call
in case application is explicitly disabling the
attribute creation order tracking with the NC_NOATTRCREORD flag.