Refactor Z library detection
HDF5 can depend on the Z library (in fact required for netCDF).  Moved the detection of whether hdf5 was built with zlib up before any other tests that may require linking of the hdf5 library to determine presence/absence of symbols.  These tests require that the link line include "-lz" if the hdf5 library was built with libz support.  

This is typically handled somewhat automatically if shared libraries are being used, but in the static library case, the explicit dependency needs to be specified.  For internal CMake checks, it uses the `CMAKE_REQUIRED_LIBRARIES` list to specify the libraries that should be used in a `CHECK_C_SOURCE_COMPILE` or a `CHECK_LIBRARY_EXISTS` call.   In the current CMakeLists.txt ordering, the zlib detection is done _after_ the `CHECK_LIBRARY_EXISTS` calls which can cause them to fail and give an incorrect result about whether the function being tested for exists.   With the reordering in this PR, I am able to correctly configure netCDF on a CRAY HPC system that uses static libraries by default.