Update selected documentation
Update the following documentation files:

## docs/FAQ.md
* Discuss the use of UTF-8 names under Windows 10+.

## docs/filters.md
* Add documentation about NCzarr filters.
  * Specifically Codec support and HDF5 <-> Codec translation
* Add documentation about standard filters

## docs/dispatch.md
* Convert from .dox format to .md (markdown) format.
* Add discussion about the user defined dispatch tables.
* Update the example.
* Abbreviate the NC_infermodel documentation and move the more detailed discusion to the companion *dinternal.md* documenation.

## docs/internal.md

This is a (mostly) new file that attempts to provide detailed
descriptions about how various features are implemented inside
the netcdf-c library. The topics currently covered the
following.

### Including C++ Code in the netcdf-c Library {#intern_c++}

The state of C compiler technology has reached the point where
it is possible to include C++ code into the netcdf-c library
code base. The document describes how to do this.

### Managing instances of complex data types

The document describes how to properly handle instances of
complex types (those with variable length).  This involves
having functions that can recursively walk instances of such
types to perform various actions on them.  These new functions
are intended to replace the *nc_free_vlen*, *nc_free_vlens* and
*nc_free_string* functions in *netcdf.h*.

### Inferring File Types

As described in the companion document -- docs/dispatch.md --
when *nc\_create()* or *nc\_open()* is called, the library must
figure out what kind of file is being created or opened.  Once it
has figured out the file kind, the appropriate "dispatch table"
can be used to process that file.

As a result of the introduction of remote data access to the netcdf-c
library, the path arguments to *nc\_open()* and *nc\_create()* have
been extended to support URLs as paths. Processing URLs requires
some significant changes to the file inference algorithm. The
details of that processing are recorded in the document.