Addendum: * If --disable-plugin is set, then force --without-plugin-dir * Propagate changes to CMakeLists.txt * Fix a couple of memory leaks in nc4internal.c and dhttp.c
I was unable to figure out a way to prevent any attempt to
install the shared libraries when --without-plugin-dir is used.
It turns out you can conditionally set the install library, but
it seems to ignore the conditional so in that case plugindir
needs to be set to some internal location in the build tree.

In any case, you still need a location for use with -rpath in
order to force creation of the .so files.