TST: Mark tst_filter.sh XFAIL on MinGW
Attempting to add a filter to a netCDF file using `nccopy` results in
a file without `_Filter` attribute set.  I suspect this is a problem
with HDF5, but I don't remember why and never got around to testing
that.  The rest of the plugin tests pass, so I'm telling the test
runner to expect this one test to fail and see if there are other
failures.