Add EXPERIMENTAL thread-safe execution support to netcdf-c
This PR is **highly experimental** and should not (yet)
be used in a production situation.

It implements thread-safety with a mechanism similar to that used by HDF5.
There is a single global mutex object that controls access to the
netcdf API. It effectively serializes all calls to the API.
For \*nix\* operating systems, it uses *pthreads* and more specifically
the *pthread_mutex* functionality. For Windows
it uses the built-in CriticalSection API.

The controlling option is *"--enable-threadsafe"* in Automake
and *"-DENABLE_THREADSAFE=on"* in CMake. In this PR, it is
enabled by default, but if it ever moves into production, it
will be disabled by default.

## Github Actions
Local build and test of this PR is known to work on
Ubuntu-21 and Visual Studio. Build and test on Github
Actions is currently failing for all systems. Fixing
will require interactive access to the relevant platform.

## Testing
This PR has had minimal testing; I am hoping that others can help out.
There is one test case included: nc_test/run_threads.sh, which uses
nc_test/tst_threads.c.

The primary functionality is in libdispatch/dmutex.c. It has
some debug flags enabled to match lock with unlock to detect
paths that bypass locking.

## Warnings
* This implementation depends heavily on using recursive mutex objects.

## Documentation
For more detailed documenation see netcdf-c/docs/threadsafe.md