BLD: Declare nulldup backup definition static not extern
The use of this function currently runs into problems with multiple definitions: once for each file including ncconfigure.h.  Defining this as static rather than extern should hide the definitions from each other.

static inline would still be closer to the definition as a macro, but that requires a #define to work on all platforms (not all compilers have inline yet).